Upload

+++ src/main/java/ru/kata/spring/boot_security/demo/services/DocumentService.java
@@ -1,9 +1,7 @@
 package ru.kata.spring.boot_security.demo.services;
 
 import ru.kata.spring.boot_security.demo.models.Document;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,7 +10,9 @@ public interface DocumentService {
 
     Document save(Document document);
 
-    Optional<Document> findById(int id);
+    Optional<Document> findById(Long id);
 
-    void delete(int id);
+    void delete(Long id);
+
+
 }
